// generated/api/front/models/NotificationType.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum NotificationType {
  ESTIMATION_REQUESTED = 'ESTIMATION_REQUESTED',
  QUOTATION_APPROVED = 'QUOTATION_APPROVED',
  QUOTATION_REJECTED = 'QUOTATION_REJECTED',
  SELLER_APPROVED = 'SELLER_APPROVED',
  SELLER_REJECTED = 'SELLER_REJECTED',
  COMMENT_ADDED = 'COMMENT_ADDED',
  WIKI_OPINION = 'WIKI_OPINION',
  WIKI_EDITED = 'WIKI_EDITED',
  WIKI_DELETED = 'WIKI_DELETED',
  WIKI_COMMENT_ADDED = 'WIKI_COMMENT_ADDED',
}